if get empty string from curl, then continue

+++ server/update.sh
@@ -6,6 +6,11 @@ cd $SLIPPER_HOME
 
 address=$(curl ifconfig.me)
 
+if [ -e $address ""]; then
+    echo "null"
+    exit 1
+fi
+
 cat << EOF > addr.js
 window.slipper = {
     "addr": "$address"
